Portability fix: Build where sizeof is not compatible with int
On Debian amd64, for example:

 util.c: In function 'iaddr_to_string':
 util.c:575:3: error: field precision specifier '.*' expects argument
 of type 'int', but argument 3 has type 'long unsigned int'

Thanks to Jonathan Amery for the report.

+++ util.c
@@ -571,7 +571,7 @@ const char *iaddr_to_string(const union iaddr *ia)
     if (r) {
 	const char fmt[]= "scoped IPv6 addr, error: %.*s";
 	sprintf(addrbuf, fmt,
-		ADNS_ADDR2TEXT_BUFLEN - sizeof(fmt) /* underestimate */,
+		(int)(ADNS_ADDR2TEXT_BUFLEN - sizeof(fmt)) /* underestimate */,
 		strerror(r));
     }
